requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.10.4
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.12.2
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
deep-translator==1.11.4
fastapi==0.95.0
frozenlist==1.3.3
googletrans==4.0.0rc1
googletranslate-python==1.0.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
magic-filter==1.0.9
markdown-it-py==3.0.0
mdurl==0.1.2
midjourney-api==1.0.7
multidict==6.0.4
openai==0.27.10
packaging==23.0
pydantic==1.10.7
Pygments==2.16.1
python-multipart==0.0.6
pytz==2023.3
replicate==0.7.0
requests==2.28.2
rfc3986==1.5.0
rich==13.6.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.26.1
tabulate==0.9.0
tqdm==4.66.1
typing_extensions==4.5.0
tzlocal==5.0.1
urllib3==1.26.15
uvicorn==0.21.1
yarl==1.8.2
